Finalization of MoveOnly support in channels

+++ test/channel_join_tests.cpp
@@ -225,7 +225,7 @@ BOOST_FIXTURE_TEST_CASE(int_join_channel_different_type_void_functor, channel_ty
     std::atomic_int result{ 0 };
 
     auto check = join(default_scheduler(),
-                      [&_result = result](int x, std::string y) { _result +=  2 + y.size(); },
+                      [&_result = result](int x, std::string y) { _result +=  2 + static_cast<int>(y.size()); },
                       receive<0>(),
                       receive<1>());
 
@@ -245,7 +245,7 @@ BOOST_FIXTURE_TEST_CASE(int_join_channel_2_join_different_type_void_functor_asyn
     std::atomic_int result{ 0 };
 
     auto check = join(default_scheduler(),
-                      [&_result = result](int x, std::string y) { _result +=  2 + y.size(); },
+                      [&_result = result](int x, std::string y) { _result +=  2 + static_cast<int>(y.size()); },
                       receive<0>(),
                       receive<1>());
